Apply standard types to sparkbarchart and sparkline chart
