+++ packages/polaris-viz-core/src/index.ts
@@ -127,8 +127,6 @@ export type {
   GradientStop,
   LineStyle,
   PartialTheme,
-  SparkBarChartProps,
-  SparkLineChartProps,
   SvgComponents,
   Theme,
   DataPoint,
