+++ packages/polaris-viz-native/src/components/SparkLineChart/stories/SparkLineChart.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import type {Story, Meta} from '@storybook/react';
 
 import {SparkLineChart} from '../';
-import type {SparkLineChartProps} from '@shopify/polaris-viz-core';
+import type {SparkLineChartProps} from '../SparkLineChart';
 
 const THEME_CONTROL_ARGS = {
   description: 'The theme that the chart will inherit its styles from',
