+++ packages/polaris-viz/src/components/SparkBarChart/SparkBarChart.tsx
@@ -1,12 +1,17 @@
-import {
-  DEFAULT_CHART_PROPS,
-  SparkBarChartProps,
-} from '@shopify/polaris-viz-core';
 import React from 'react';
+import {DEFAULT_CHART_PROPS} from '@shopify/polaris-viz-core';
+import type {Dimensions, ChartProps} from '@shopify/polaris-viz-core';
 
 import {ChartContainer} from '../ChartContainer';
 
 import {Chart} from './Chart';
+
+export type SparkBarChartProps = {
+  dataOffsetRight?: number;
+  dataOffsetLeft?: number;
+  accessibilityLabel?: string;
+  dimensions?: Dimensions;
+} & ChartProps;
 
 export function SparkBarChart(props: SparkBarChartProps) {
   const {
